+++ src/Service/InstallationCommandsRunner.php
@@ -7,6 +7,7 @@ use Symfony\Component\Console\Helper\ProcessHelper;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\Process\PhpExecutableFinder;
+
 use function array_unshift;
 use function implode;
 use function sprintf;