+++ test/Config/Plugin/ApplicationConfigCustomizerTest.php
@@ -12,6 +12,7 @@ use Shlinkio\Shlink\Installer\Model\CustomizableAppConfig;
 use Shlinkio\Shlink\Installer\Util\StringGeneratorInterface;
 use ShlinkioTest\Shlink\Installer\Util\TestUtilsTrait;
 use Symfony\Component\Console\Style\SymfonyStyle;
+
 use function array_shift;
 use function strpos;