+++ src/Config/Util/ExpectedConfigResolverFactory.php
@@ -8,7 +8,7 @@ use Zend\ServiceManager\Factory\FactoryInterface;
 
 class ExpectedConfigResolverFactory implements FactoryInterface
 {
-    public function __invoke(ContainerInterface $container, $requestedName, array $options = null)
+    public function __invoke(ContainerInterface $container, $requestedName, ?array $options = null)
     {
         $expectedConfigMap = $container->get('config')['installer_plugins_expected_config'] ?? [];
         return new ExpectedConfigResolver($expectedConfigMap);