+++ config/config.php
@@ -3,8 +3,10 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer;
 
-use Symfony\Component\Console\Application;
+use Shlinkio\Shlink\Installer\Factory\ProcessHelperFactory;
+use Symfony\Component\Console;
 use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Process\PhpExecutableFinder;
 use Zend\ServiceManager\AbstractFactory\ConfigAbstractFactory;
 use Zend\ServiceManager\Factory\InvokableFactory;
 
@@ -12,8 +14,11 @@ return [
 
     'dependencies' => [
         'factories' => [
-            Application::class => Factory\InstallApplicationFactory::class,
+            Console\Application::class => Factory\InstallApplicationFactory::class,
             Filesystem::class => InvokableFactory::class,
+            PhpExecutableFinder::class => InvokableFactory::class,
+            Console\Helper\ProcessHelper::class => ProcessHelperFactory::class,
+
             Util\StringGenerator::class => InvokableFactory::class,
             Service\InstallationCommandsRunner::class => Service\InstallationCommandsRunnerFactory::class,
             Config\Util\ExpectedConfigResolver::class => Config\Util\ExpectedConfigResolverFactory::class,
@@ -44,22 +49,22 @@ return [
 
     'installation_commands' => [
         'db_create_schema' => [
-            'command' => ['vendor/doctrine/orm/bin/doctrine.php', 'orm:schema-tool:create'],
+            'command' => 'vendor/doctrine/orm/bin/doctrine.php orm:schema-tool:create',
             'initMessage' => 'Initializing database...',
             'errorMessage' => 'Error generating database.',
         ],
         'db_migrate' => [
-            'command' => ['vendor/doctrine/migrations/bin/doctrine-migrations.php', 'migrations:migrate'],
+            'command' => 'vendor/doctrine/migrations/bin/doctrine-migrations.php migrations:migrate',
             'initMessage' => 'Updating database...',
             'errorMessage' => 'Error updating database.',
         ],
         'orm_proxies' => [
-            'command' => ['vendor/doctrine/orm/bin/doctrine.php', 'orm:generate-proxies'],
+            'command' => 'vendor/doctrine/orm/bin/doctrine.php orm:generate-proxies',
             'initMessage' => 'Generating proxies...',
             'errorMessage' => 'Error generating proxies.',
         ],
         'geolite_download' => [
-            'command' => ['bin/cli', 'visit:update-db'],
+            'command' => 'bin/cli visit:update-db',
             'initMessage' => 'Downloading GeoLite2 db...',
             'errorMessage' => 'Error downloading GeoLite2 db.',
         ],