+++ CHANGELOG.md
@@ -4,11 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
-## [Unreleased]
+## 1.2.0 - 2019-08-03
 
 #### Added
 
-* *Nothing*
+* [#17](https://github.com/shlinkio/shlink-installer/issues/17) Allowed commands run during installation to be overwritten via configuration.
 
 #### Changed